# backend/docs/swagger.yaml
basePath: /api
definitions:
...
  furniture-commission_internal_service.CreateUserRequest:
    properties:
      bank_account:
        example: "6222021234567890123"
        type: string
      bank_name:
        example: 中国工商银行
        type: string
      base_salary:
        type: number
      department_id:
        type: integer
      email:
        example: zhangsan@example.com
        type: string
      employee_no:
        example: EMP001
        type: string
      entry_date:
        example: "2024-01-01"
        type: string
      id_card:
        example: "110101199001011234"
        type: string
      is_formal:
        type: integer
      parent_id:
        type: integer
      password:
        example: "123456"
        minLength: 6
        type: string
      phone:
        example: "13800138000"
        type: string
      probation_end_date:
        example: "2024-04-01"
        type: string
      real_name:
        example: 张三
        type: string
      referrer_id:
        type: integer
      role:
        type: integer
      role_ids:
        items:
          type: integer
        type: array
      store_id:
        type: integer
      username:
        example: zhangsan
        type: string
    required:
    - password
    - username
    type: object
...
  furniture-commission_internal_service.UpdateUserRequest:
    properties:
      bank_account:
        example: "6222021234567890123"
        type: string
      bank_name:
        example: 中国工商银行
        type: string
      base_salary:
        type: number
      department_id:
        type: integer
      email:
        example: zhangsan@example.com
        type: string
      employee_no:
        example: EMP001
        type: string
      entry_date:
        example: "2024-01-01"
        type: string
      id_card:
        example: "110101199001011234"
        type: string
      is_formal:
        type: integer
      parent_id:
        type: integer
      phone:
        example: "13800138000"
        type: string
      probation_end_date:
        example: "2024-04-01"
        type: string
      real_name:
        example: 张三
        type: string
      referrer_id:
        type: integer
      role:
        type: integer
      status:
        type: integer
      store_id:
        type: integer
    type: object
...
  internal_handler.AssignRoleRequest:
    properties:
      role_ids:
        example:
        - 1
        - 2
        - 3
        items:
          type: integer
        type: array
    required:
    - role_ids
    type: object
...
  internal_handler.Response:
    properties:
      code:
        example: 200
        type: integer
      data: {}
      message:
        example: success
        type: string
      timestamp:
        example: 1715000000
        type: integer
    type: object
...
  internal_handler.UpdateUserStatusRequest:
    properties:
      status:
        example: 1
        type: integer
    type: object
...
host: localhost:8080
info:
  contact:
    email: support@example.com
    name: API 支持
  description: 家具提成管理系统后端 API 接口文档
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: 家具提成管理系统 API
  version: "1.0"
paths:
...
  /users:
    get:
      consumes:
      - application/json
      description: 分页查询用户列表，支持按关键词、角色、状态筛选
      parameters:
      - default: 1
        description: 页码
        in: query
        name: page
        type: integer
      - default: 10
        description: 每页数量
        in: query
        name: page_size
        type: integer
      - description: 搜索关键词
        in: query
        name: keyword
        type: string
      - description: 角色筛选
        in: query
        name: role
        type: integer
      - description: 状态筛选
        in: query
        name: status
        type: integer
      - description: 门店ID
        in: query
        name: store_id
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: 业务错误
          schema:
            $ref: '#/definitions/internal_handler.Response'
        "401":
          description: 未认证
          schema:
            $ref: '#/definitions/internal_handler.Response'
      security:
      - BearerAuth: []
      summary: 获取用户列表
      tags:
      - 用户管理
    post:
      consumes:
      - application/json
      description: 创建新用户，需要管理员权限
      parameters:
      - description: 创建用户请求
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/furniture-commission_internal_service.CreateUserRequest'
      produces:
      - application/json
      responses:
        "200":
          description: 业务错误
          schema:
            $ref: '#/definitions/internal_handler.Response'
        "401":
          description: 未认证
          schema:
            $ref: '#/definitions/internal_handler.Response'
      security:
      - BearerAuth: []
      summary: 创建用户
      tags:
      - 用户管理
  /users/{id}:
    delete:
      consumes:
      - application/json
      description: 根据用户ID删除用户，需要管理员权限
      parameters:
      - description: 用户ID
        format: int64
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: 业务错误
          schema:
            $ref: '#/definitions/internal_handler.Response'
        "401":
          description: 未认证
          schema:
            $ref: '#/definitions/internal_handler.Response'
      security:
      - BearerAuth: []
      summary: 删除用户
      tags:
      - 用户管理
    get:
      consumes:
      - application/json
      description: 根据用户ID获取用户详细信息
      parameters:
      - description: 用户ID
        format: int64
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: 业务错误
          schema:
            $ref: '#/definitions/internal_handler.Response'
        "401":
          description: 未认证
          schema:
            $ref: '#/definitions/internal_handler.Response'
      security:
      - BearerAuth: []
      summary: 获取用户详情
      tags:
      - 用户管理
    put:
      consumes:
      - application/json
      description: 根据用户ID更新用户信息
      parameters:
      - description: 用户ID
        format: int64
        in: path
        name: id
        required: true
        type: integer
      - description: 更新用户请求
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/furniture-commission_internal_service.UpdateUserRequest'
      produces:
      - application/json
      responses:
        "200":
          description: 业务错误
          schema:
            $ref: '#/definitions/internal_handler.Response'
        "401":
          description: 未认证
          schema:
            $ref: '#/definitions/internal_handler.Response'
      security:
      - BearerAuth: []
      summary: 更新用户
      tags:
      - 用户管理
  /users/{id}/reset-password:
    post:
      consumes:
      - application/json
      description: 重置指定用户的密码，返回新密码，需要管理员权限
      parameters:
      - description: 用户ID
        format: int64
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: 业务错误
          schema:
            $ref: '#/definitions/internal_handler.Response'
        "401":
          description: 未认证
          schema:
            $ref: '#/definitions/internal_handler.Response'
      security:
      - BearerAuth: []
      summary: 重置密码
      tags:
      - 用户管理
  /users/{id}/roles:
    post:
      consumes:
      - application/json
      description: 为用户分配一个或多个角色，需要管理员权限
      parameters:
      - description: 用户ID
        format: int64
        in: path
        name: id
        required: true
        type: integer
      - description: 分配角色请求
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/internal_handler.AssignRoleRequest'
      produces:
      - application/json
      responses:
        "200":
          description: 业务错误
          schema:
            $ref: '#/definitions/internal_handler.Response'
        "401":
          description: 未认证
          schema:
            $ref: '#/definitions/internal_handler.Response'
      security:
      - BearerAuth: []
      summary: 分配角色
      tags:
      - 用户管理
  /users/{id}/status:
    put:
      consumes:
      - application/json
      description: 启用或禁用用户账号，需要管理员权限
      parameters:
      - description: 用户ID
        format: int64
        in: path
        name: id
        required: true
        type: integer
      - description: 状态更新请求
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/internal_handler.UpdateUserStatusRequest'
      produces:
      - application/json
      responses:
        "200":
          description: 业务错误
          schema:
            $ref: '#/definitions/internal_handler.Response'
        "401":
          description: 未认证
          schema:
            $ref: '#/definitions/internal_handler.Response'
      security:
      - BearerAuth: []
      summary: 更新用户状态
      tags:
      - 用户管理
...
securityDefinitions:
  BearerAuth:
    description: 'JWT Bearer Token 认证，格式: Bearer {token}'
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"